+++ enn_zoo/enn_zoo/train.py
@@ -14,8 +14,8 @@ from enn_zoo.griddly_env import GRIDDLY_ENVS
 from enn_zoo.codecraft.cc_vec_env import codecraft_env_class, CodeCraftVecEnv
 from enn_zoo.codecraft.codecraftnet.adapter import CCNetAdapter
 from enn_zoo.microrts import GymMicrorts
-from enn_zoo import vizdoom_envs
-from enn_zoo.vizdoom_envs import VIZDOOM_ENVS
+from enn_zoo import vizdoom_env
+from enn_zoo.vizdoom_env import VIZDOOM_ENVS
 
 
 @dataclass
@@ -62,7 +62,7 @@ def main(cfg: TrainConfig) -> None:
     elif cfg.env.id == "GymMicrorts":
         env_cls = GymMicrorts
     elif cfg.env.id in VIZDOOM_ENVS:
-        env_cls = vizdoom_envs.create_vizdoom_env(VIZDOOM_ENVS[cfg.env.id])
+        env_cls = vizdoom_env.create_vizdoom_env(VIZDOOM_ENVS[cfg.env.id])
     else:
         raise KeyError(
             f"Unknown gym_id: {cfg.env.id}\nAvailable environments: {list(ENV_REGISTRY.keys()) + list(GRIDDLY_ENVS.keys()) + ['CodeCraft']}"
