+++ enn_zoo/enn_zoo/vizdoom_env/scenarios/basic.wad
