+++ enn_zoo/enn_zoo/vizdoom_env/scenarios/defend_the_center.cfg
