+++ enn_zoo/enn_zoo/vizdoom_env/scenarios/health_gathering.cfg
